Extract Excel worksheet helpers
